made a file for pending user skills endorsements as well and removed from gensec endorsement page
